+++ src/main/java/com/jeontongju/consumer/dto/temp/TradePathEnum.java
@@ -1,4 +1,4 @@
-package com.jeontongju.consumer.enums;
+package com.jeontongju.consumer.dto.temp;
 
 import lombok.Getter;
